+++ Gameplay/Items/Resources/Cloth.cs
@@ -4,7 +4,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "Cotton";
         public override string Name => "Cotton";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 1;
     }
 
@@ -12,7 +12,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "Fiber";
         public override string Name => "Fiber";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 1;
     }
 
@@ -20,7 +20,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "ArcaneFiber";
         public override string Name => "Arcane Fiber";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 3;
         public override ItemRarity Rarity => ItemRarity.Uncommon;
     }
@@ -29,7 +29,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "MasterCotton";
         public override string Name => "Master Cotton";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 5;
         public override ItemRarity Rarity => ItemRarity.Uncommon;
     }
@@ -38,7 +38,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "Silk";
         public override string Name => "Silk";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
         public override ItemRarity Rarity => ItemRarity.Rare;
     }
@@ -47,7 +47,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "MysticSilk";
         public override string Name => "Mystic Silk";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 25;
         public override ItemRarity Rarity => ItemRarity.Rare;
     }
@@ -56,7 +56,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DragonFlowerSilk";
         public override string Name => "Dragon Flower Silk";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 30;
         public override ItemRarity Rarity => ItemRarity.Magic;
     }
@@ -65,7 +65,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DivineCotton";
         public override string Name => "Divine Cotton";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 100;
         public override ItemRarity Rarity => ItemRarity.Legendary;
     }
@@ -75,7 +75,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "LinenCloth";
         public override string Name => "Linen Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 1;
     }
 
@@ -83,7 +83,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "WoolenCloth";
         public override string Name => "Woolen Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 1;
     }
 
@@ -91,7 +91,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "MageweaveCloth";
         public override string Name => "Mageweave Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 3;
         public override ItemRarity Rarity => ItemRarity.Uncommon;
     }
@@ -100,7 +100,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "PrimalCloth";
         public override string Name => "Primal Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 5;
         public override ItemRarity Rarity => ItemRarity.Uncommon;
     }
@@ -109,7 +109,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "MagicCloth";
         public override string Name => "Magic Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
         public override ItemRarity Rarity => ItemRarity.Rare;
     }
@@ -118,7 +118,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "SilkCloth";
         public override string Name => "Silk Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
         public override ItemRarity Rarity => ItemRarity.Rare;
     }
@@ -127,7 +127,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "EnchantedCloth";
         public override string Name => "Enchanted Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 15;
         public override ItemRarity Rarity => ItemRarity.Magic;
     }
@@ -136,7 +136,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DemonicCloth";
         public override string Name => "Demonic Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 30;
         public override ItemRarity Rarity => ItemRarity.Magic;
     }
@@ -145,7 +145,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DarknessCloth";
         public override string Name => "Darkness Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 35;
         public override ItemRarity Rarity => ItemRarity.Magic;
     }
@@ -154,7 +154,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DragonestCloth";
         public override string Name => "Dragonest Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 50;
         public override ItemRarity Rarity => ItemRarity.Legendary;
     }
@@ -163,7 +163,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DivineCloth";
         public override string Name => "Divine Cloth";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 100;
         public override ItemRarity Rarity => ItemRarity.Legendary;
     }
