+++ Gameplay/Items/Resources/Enchantments.cs
@@ -4,7 +4,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "ColdEssence";
         public override string Name => "Cold Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
     }
 
@@ -12,7 +12,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "DarknessEssence";
         public override string Name => "Darkness Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 100;
     }
 
@@ -20,7 +20,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "EarthEssence";
         public override string Name => "Earth Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
     }
 
@@ -28,7 +28,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "FireEssence";
         public override string Name => "Fire Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
     }
 
@@ -36,7 +36,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "LightEssence";
         public override string Name => "Light Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
     }
 
@@ -44,7 +44,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "NatureEssence";
         public override string Name => "Nature Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
     }
 
@@ -52,7 +52,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "WindEssence";
         public override string Name => "Wind Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 10;
     }
 
@@ -60,7 +60,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "ElementalDust";
         public override string Name => "Elemental Dust";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 50;
     }
 
@@ -68,7 +68,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "MagicDust";
         public override string Name => "Magic Dust";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 100;
     }
 
@@ -76,7 +76,7 @@ namespace Server.Gameplay.Items
     {
         public override string Namespace => "MagicEssence";
         public override string Name => "Magic Essence";
-        public override double Weight => 0.1;
+        public override float Weight => 0.1f;
         public override int GoldCost => 300;
     }
 }
